Rename files and traits
