+++ application/comit_node/src/swap_protocols/rfc003/actions/mod.rs
@@ -1,7 +1,11 @@
-use crate::swap_protocols::rfc003::Timestamp;
+use crate::swap_protocols::{
+    asset::Asset,
+    rfc003::{secret_source::SecretSource, state_machine::HtlcParams, Ledger, Secret, Timestamp},
+};
 
+pub mod bitcoin;
 pub mod erc20;
-pub mod non_erc20;
+pub mod ether;
 
 pub trait Actions {
     type ActionKind;
@@ -22,4 +26,25 @@ impl<ActionKind> Action<ActionKind> {
             ..self
         }
     }
+}
+
+pub trait OneStepFundActions<L: Ledger, A: Asset> {
+    type FundActionOutput;
+    type RefundActionOutput;
+    type RedeemActionOutput;
+
+    fn fund_action(htlc_params: HtlcParams<L, A>) -> Self::FundActionOutput;
+
+    fn refund_action(
+        htlc_params: HtlcParams<L, A>,
+        htlc_location: L::HtlcLocation,
+        secret_source: &dyn SecretSource,
+    ) -> Self::RefundActionOutput;
+
+    fn redeem_action(
+        htlc_params: HtlcParams<L, A>,
+        htlc_location: L::HtlcLocation,
+        secret_source: &dyn SecretSource,
+        secret: Secret,
+    ) -> Self::RedeemActionOutput;
 }
