+++ application/comit_node/src/swap_protocols/rfc003/alice/actions/mod.rs
@@ -1,7 +1,7 @@
 use crate::swap_protocols::rfc003::actions::Action;
 
 mod erc20;
-mod non_erc20;
+mod one_step_fund_actions;
 
 #[derive(Debug, Clone, PartialOrd, Ord, PartialEq, Eq, Hash)]
 pub enum ActionKind<Deploy, Fund, Redeem, Refund> {
