+++ application/comit_node/src/swap_protocols/rfc003/bob/actions/one_step_fund_actions.rs
@@ -1,7 +1,7 @@
 use crate::swap_protocols::{
     asset::Asset,
     rfc003::{
-        actions::{non_erc20::CreateActions, Actions},
+        actions::{Actions, OneStepFundActions},
         bob::{
             self,
             actions::{Accept, Decline},
@@ -19,17 +19,17 @@ where
     BL: Ledger,
     AA: Asset,
     BA: Asset,
-    (AL, AA): CreateActions<AL, AA>,
-    (BL, BA): CreateActions<BL, BA>,
+    (AL, AA): OneStepFundActions<AL, AA>,
+    (BL, BA): OneStepFundActions<BL, BA>,
 {
     #[allow(clippy::type_complexity)]
     type ActionKind = bob::ActionKind<
         Accept<AL, BL>,
         Decline<AL, BL>,
         (),
-        <(BL, BA) as CreateActions<BL, BA>>::FundActionOutput,
-        <(AL, AA) as CreateActions<AL, AA>>::RedeemActionOutput,
-        <(BL, BA) as CreateActions<BL, BA>>::RefundActionOutput,
+        <(BL, BA) as OneStepFundActions<BL, BA>>::FundActionOutput,
+        <(AL, AA) as OneStepFundActions<AL, AA>>::RedeemActionOutput,
+        <(BL, BA) as OneStepFundActions<BL, BA>>::RefundActionOutput,
     >;
 
     fn actions(&self) -> Vec<Action<Self::ActionKind>> {
